add post image id
