+++ scripts/entry.js
@@ -4,4 +4,5 @@ require("./merge-config").mergeConfig(hexo);
 require("./lazyload").lazyload(hexo);
 require("./title").title(hexo);
 require("./carrier").carrier(hexo);
-require("./random-image").randomImage(hexo);
+require("./post-image").randomImage(hexo);
+require("./post-image").getImageById(hexo);
